squash: rename `ns.h` -> `namespace.h`
