+++ src/openassetio-core-c/include/openassetio/c/managerAPI/ManagerInterface.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "../SimpleString.h"
-#include "../ns.h"
+#include "../namespace.h"
 
 #ifdef __cplusplus
 extern "C" {
