+++ src/openassetio-core-c/include/openassetio/c/namespace.h
